Extract MAUI-free Microsoft.Maui.DevFlow.Agent.Abstractions
DevFlow's agent could only ever run inside a MAUI app because
Microsoft.Maui.DevFlow.Agent.Core hard-references Microsoft.Maui.Controls
and Microsoft.Maui.Essentials. Split the framework-neutral half out into a
new net10.0 assembly with zero MAUI references so plain .NET Android, iOS,
Mac Catalyst and macOS apps can host the same agent.

Microsoft.Maui.DevFlow.Agent.Abstractions now owns the HTTP server, routing,
CSS engine, network capture, profiler, extensions, invoke/actions, file
storage plumbing and the DevFlowAgentService routing core. Every endpoint
that needs a UI framework is a protected virtual handler that returns a
uniform 501 { error: "not_supported", capability, reason } envelope by
default. Agent.Core keeps its PackageId and becomes the MAUI backend:
MauiDevFlowAgentService : DevFlowAgentService overrides those handlers with
the existing bodies verbatim, so Agent, Agent.Gtk and Agent.WPF keep every
protected virtual hook they already override.

Three seams replace direct MAUI dependencies in the neutral half:
IAgentDispatcher replaces IDispatcher, a static FrameworkValueFormatter hook
replaces MAUI-aware property formatting, and RuntimeProfilerCollector takes a
DisplayRefreshRateProvider delegate instead of calling DeviceDisplay. Essentials
exception types are matched by name-walking rather than by type.

Types keep their existing namespaces, so the move is source-compatible but
binary-breaking for anyone who referenced them directly from Agent.Core.

Copilot-Session: 168e7050-987f-4782-9527-a8fa40e67bdd
