+++ MauiLabs.slnx
@@ -42,6 +42,7 @@
     <Project Path="src/Cli/Microsoft.Maui.ProfilingHelper/Microsoft.Maui.ProfilingHelper.csproj" />
   </Folder>
   <Folder Name="/src/DevFlow/">
+    <Project Path="src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Microsoft.Maui.DevFlow.Agent.Abstractions.csproj" />
     <Project Path="src/DevFlow/Microsoft.Maui.DevFlow.Agent.Core/Microsoft.Maui.DevFlow.Agent.Core.csproj" />
     <Project Path="src/DevFlow/Microsoft.Maui.DevFlow.Analyzers/Microsoft.Maui.DevFlow.Analyzers.csproj" />
     <Project Path="src/DevFlow/Microsoft.Maui.DevFlow.Agent.Gtk/Microsoft.Maui.DevFlow.Agent.Gtk.csproj" />
