+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/AgentExtension.cs
