+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/BrokerRegistration.cs
