+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Css/ElementInfoOps.cs
