+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/ElementInfo.cs
