+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/FileStoragePathResolver.cs
