+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Microsoft.Maui.DevFlow.Agent.Abstractions.csproj
@@ -1,0 +1,25 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <IsPackable>true</IsPackable>
+    <IsShipping>true</IsShipping>
+
+    <PackageId>Microsoft.Maui.DevFlow.Agent.Abstractions</PackageId>
+    <Description>Framework-neutral agent infrastructure for Microsoft.Maui.DevFlow. HTTP server, protocol handling, CSS querying, network capture, and profiling — with no dependency on .NET MAUI. UI frameworks plug in through the backend interfaces.</Description>
+    <PackageTags>maui;automation;testing;agent;abstractions</PackageTags>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Fizzler" />
+    <PackageReference Include="SkiaSharp" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Microsoft.Maui.DevFlow.Logging\Microsoft.Maui.DevFlow.Logging.csproj" />
+    <ProjectReference Include="..\Microsoft.Maui.DevFlow.Analyzers\Microsoft.Maui.DevFlow.Analyzers.csproj"
+                      ReferenceOutputAssembly="false"
+                      OutputItemType="Analyzer" />
+  </ItemGroup>
+
+</Project>
