+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Network/NetworkRequestEntry.cs
