+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Network/NetworkRequestStore.cs
