+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Profiling/IMarkerPublisher.cs
