+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Profiling/INativeFrameStatsProvider.cs
