+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Profiling/IProfilerCollector.cs
