+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Profiling/ProfilerContracts.cs
