+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Profiling/ProfilerRingBuffer.cs
