+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Profiling/ProfilerSessionStore.cs
