+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Abstractions/Properties/InternalsVisibleTo.cs
@@ -1,0 +1,8 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Microsoft.Maui.DevFlow.Tests")]
+[assembly: InternalsVisibleTo("Microsoft.Maui.DevFlow.Agent.Core")]
+[assembly: InternalsVisibleTo("Microsoft.Maui.DevFlow.Agent.Native")]
+[assembly: InternalsVisibleTo("Microsoft.Maui.DevFlow.Agent")]
+[assembly: InternalsVisibleTo("Microsoft.Maui.DevFlow.Agent.Gtk")]
+[assembly: InternalsVisibleTo("Microsoft.Maui.DevFlow.Agent.WPF")]
