+++ src/DevFlow/Microsoft.Maui.DevFlow.Agent.Core/Microsoft.Maui.DevFlow.Agent.Core.csproj
@@ -6,7 +6,7 @@
     <IsShipping>true</IsShipping>
 
     <PackageId>Microsoft.Maui.DevFlow.Agent.Core</PackageId>
-    <Description>Core agent infrastructure for Microsoft.Maui.DevFlow. HTTP server, visual tree walking, logging, and element interaction — platform-agnostic.</Description>
+    <Description>The .NET MAUI backend for Microsoft.Maui.DevFlow. Implements visual tree walking, element interaction, screenshots, storage and device endpoints on top of Microsoft.Maui.DevFlow.Agent.Abstractions.</Description>
     <PackageTags>maui;automation;testing;agent;core</PackageTags>
   </PropertyGroup>
 
@@ -18,6 +18,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Microsoft.Maui.DevFlow.Agent.Abstractions\Microsoft.Maui.DevFlow.Agent.Abstractions.csproj" />
     <ProjectReference Include="..\Microsoft.Maui.DevFlow.Logging\Microsoft.Maui.DevFlow.Logging.csproj" />
     <ProjectReference Include="..\Microsoft.Maui.DevFlow.Analyzers\Microsoft.Maui.DevFlow.Analyzers.csproj"
                       ReferenceOutputAssembly="false"
